Change doller.times()
I just noticed that in the Java code of the book, times() is a
static method; to define a static method in Python, use the
decorator `@staticmethod`. Also, if the return value of a function
is nothing, i.e., void in Java, you can use the `None` type
annotation in Python.